Add files via upload
Data Manipulation with Pandas is a Data camp course of a Data science roadmap. In this course, I have learned data manipulating techniques with Python where we have to sort, index, slice, aggregate, and so on to make data more efficient and readable. In this code, I have copied all the exercise provided by this course.